fixing link error

+++ src/vectors/trilinos/tpetra/TpetraVectorData.h
@@ -15,7 +15,7 @@ namespace AMP::LinearAlgebra {
 
 template<typename ST = double,
          typename LO = int32_t,
-         typename GO = int64_t,
+         typename GO = long long,
          typename NT = Tpetra::Vector<>::node_type>
 class TpetraVectorData : public VectorData
 {
